# Optimize evictedQueue implementation and use