--- a/sdk/trace/evictedqueue.go
+++ b/sdk/trace/evictedqueue.go
@@ -27,7 +27,7 @@ func newEvictedQueue(capacity int) evictedQueue {
 }
 
 // add adds value to the evictedQueue eq. If eq is at capacity, the oldest
-// queued value will be discarded the drop count incremented.
+// queued value will be discarded and the drop count incremented.
 func (eq *evictedQueue) add(value interface{}) {
 	if len(eq.queue) == eq.capacity {
 		// Drop first-in while avoiding allocating more capacity to eq.queue.