--- a/sdk/trace/span.go
+++ b/sdk/trace/span.go
@@ -142,10 +142,10 @@ type recordingSpan struct {
 	attributes *attributesMap
 
 	// events are stored in FIFO queue capped by configured limit.
-	events *evictedQueue
+	events evictedQueue
 
 	// links are stored in FIFO queue capped by configured limit.
-	links *evictedQueue
+	links evictedQueue
 
 	// executionTracerTaskEnd ends the execution tracer span.
 	executionTracerTaskEnd func()