Taken from performance improvements identified in https://github.com/open-telemetry/opentelemetry-go/issues/2402.

- Avoid unnecessary allocations in the `recordingSpan` by using an `evictedQueue` type instead of a pointer to one.
- Lazy allocate the `evictedQueue` queue to prevent unnecessary operations for spans without any use of the queue.
- Document the `evictedQueue`

### Testing Performance

This looks to avoid 4 allocations for spans that do not add events or links and includes memory/allocation improvements for those that do.

`cd sdk/trace && go test -bench=.`

```
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/otel/sdk/trace
cpu: Intel(R) Core(TM) i7-8550U CPU @ 1.80GHz
BenchmarkAttributesMapToKeyValue-8         	  397270	      2850 ns/op
BenchmarkSpanProcessor-8                   	   98577	     11676 ns/op	   10352 B/op	      65 allocs/op
BenchmarkSpanProcessorVerboseLogging-8     	   94104	     12617 ns/op	   11496 B/op	      71 allocs/op
BenchmarkStartEndSpan/AlwaysSample-8       	 1582981	       779.1 ns/op	     624 B/op	       5 allocs/op
BenchmarkStartEndSpan/NeverSample-8        	 4077541	       289.7 ns/op	     128 B/op	       2 allocs/op
BenchmarkSpanWithAttributes_4/AlwaysSample-8         	  701468	      1652 ns/op	    1536 B/op	      15 allocs/op
BenchmarkSpanWithAttributes_4/NeverSample-8          	 2698389	       461.1 ns/op	     384 B/op	       3 allocs/op
BenchmarkSpanWithAttributes_8/AlwaysSample-8         	  415170	      2697 ns/op	    2240 B/op	      23 allocs/op
BenchmarkSpanWithAttributes_8/NeverSample-8          	 1877380	       629.1 ns/op	     640 B/op	       3 allocs/op
BenchmarkSpanWithAttributes_all/AlwaysSample-8       	  569984	      1877 ns/op	    1712 B/op	      17 allocs/op
BenchmarkSpanWithAttributes_all/NeverSample-8        	 2605310	       493.1 ns/op	     448 B/op	       3 allocs/op
BenchmarkSpanWithAttributes_all_2x/AlwaysSample-8    	  357108	      3231 ns/op	    3012 B/op	      28 allocs/op
BenchmarkSpanWithAttributes_all_2x/NeverSample-8     	 1982054	       632.0 ns/op	     768 B/op	       3 allocs/op
BenchmarkSpanWithEvents_4/AlwaysSample-8             	  762336	      1542 ns/op	    1056 B/op	      12 allocs/op
BenchmarkSpanWithEvents_4/NeverSample-8              	 4106401	       292.6 ns/op	     128 B/op	       2 allocs/op
BenchmarkSpanWithEvents_8/AlwaysSample-8             	  531265	      2231 ns/op	    1504 B/op	      17 allocs/op
BenchmarkSpanWithEvents_8/NeverSample-8              	 3971354	       303.8 ns/op	     128 B/op	       2 allocs/op
BenchmarkSpanWithEvents_WithStackTrace/AlwaysSample-8         	 1000000	      1043 ns/op	     736 B/op	       8 allocs/op
BenchmarkSpanWithEvents_WithStackTrace/NeverSample-8          	 3605416	       327.2 ns/op	     144 B/op	       3 allocs/op
BenchmarkSpanWithEvents_WithTimestamp/AlwaysSample-8          	 1000000	      1018 ns/op	     760 B/op	       9 allocs/op
BenchmarkSpanWithEvents_WithTimestamp/NeverSample-8           	 3178768	       393.5 ns/op	     168 B/op	       4 allocs/op
BenchmarkTraceID_DotString-8                                  	15191973	        75.54 ns/op
BenchmarkSpanID_DotString-8                                   	20735128	        57.35 ns/op
PASS
```

#### Reference: [main branch](https://github.com/open-telemetry/opentelemetry-go/commit/86ced11f2c060e479817477f614438faa799ce93) benchmarks

```
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/otel/sdk/trace
cpu: Intel(R) Core(TM) i7-8550U CPU @ 1.80GHz
BenchmarkAttributesMapToKeyValue-8         	  398217	      2914 ns/op
BenchmarkSpanProcessor-8                   	   88304	     13970 ns/op	   12592 B/op	     105 allocs/op
BenchmarkSpanProcessorVerboseLogging-8     	   83248	     13659 ns/op	   13736 B/op	     111 allocs/op
BenchmarkStartEndSpan/AlwaysSample-8       	 1000000	      1000 ns/op	     848 B/op	       9 allocs/op
BenchmarkStartEndSpan/NeverSample-8        	 3449281	       368.1 ns/op	     224 B/op	       3 allocs/op
BenchmarkSpanWithAttributes_4/AlwaysSample-8         	  626064	      1811 ns/op	    1760 B/op	      19 allocs/op
BenchmarkSpanWithAttributes_4/NeverSample-8          	 2442355	       486.2 ns/op	     480 B/op	       4 allocs/op
BenchmarkSpanWithAttributes_8/AlwaysSample-8         	  449365	      2646 ns/op	    2464 B/op	      27 allocs/op
BenchmarkSpanWithAttributes_8/NeverSample-8          	 1877700	       631.6 ns/op	     736 B/op	       4 allocs/op
BenchmarkSpanWithAttributes_all/AlwaysSample-8       	  585984	      1931 ns/op	    1936 B/op	      21 allocs/op
BenchmarkSpanWithAttributes_all/NeverSample-8        	 2300470	       513.6 ns/op	     544 B/op	       4 allocs/op
BenchmarkSpanWithAttributes_all_2x/AlwaysSample-8    	  333524	      3456 ns/op	    3236 B/op	      32 allocs/op
BenchmarkSpanWithAttributes_all_2x/NeverSample-8     	 1729068	       696.2 ns/op	     864 B/op	       4 allocs/op
BenchmarkSpanWithEvents_4/AlwaysSample-8             	  577074	      1978 ns/op	    1536 B/op	      20 allocs/op
BenchmarkSpanWithEvents_4/NeverSample-8              	 3241960	       374.9 ns/op	     224 B/op	       3 allocs/op
BenchmarkSpanWithEvents_8/AlwaysSample-8             	  363082	      3137 ns/op	    2240 B/op	      29 allocs/op
BenchmarkSpanWithEvents_8/NeverSample-8              	 2901153	       384.8 ns/op	     224 B/op	       3 allocs/op
BenchmarkSpanWithEvents_WithStackTrace/AlwaysSample-8         	  866670	      1420 ns/op	    1024 B/op	      13 allocs/op
BenchmarkSpanWithEvents_WithStackTrace/NeverSample-8          	 2716603	       437.5 ns/op	     240 B/op	       4 allocs/op
BenchmarkSpanWithEvents_WithTimestamp/AlwaysSample-8          	  971718	      1337 ns/op	    1048 B/op	      14 allocs/op
BenchmarkSpanWithEvents_WithTimestamp/NeverSample-8           	 2532778	       484.3 ns/op	     264 B/op	       5 allocs/op
BenchmarkTraceID_DotString-8                                  	14741889	        80.00 ns/op
BenchmarkSpanID_DotString-8                                   	19892643	        58.77 ns/op
PASS
```