@MadVikingGod 2022-01-26
I ran the branch through benchstat to give a more detailed view of the differences.  It looks great!


```
benchstat bench-main bench-branch 
name                                          old time/op    new time/op    delta
AttributesMapToKeyValue-4                       4.68µs ±55%    4.11µs ± 7%  -12.10%  (p=0.004 n=8+10)
SpanProcessor-4                                 22.3µs ±64%    17.7µs ± 6%  -20.65%  (p=0.001 n=9+10)
SpanProcessorVerboseLogging-4                   25.4µs ±71%    18.1µs ±11%  -28.72%  (p=0.000 n=10+10)
StartEndSpan/AlwaysSample-4                     1.27µs ±21%    1.06µs ± 3%  -16.72%  (p=0.000 n=8+8)
StartEndSpan/NeverSample-4                       441ns ±17%     429ns ± 7%     ~     (p=0.650 n=8+10)
SpanWithAttributes_4/AlwaysSample-4             3.05µs ±74%    2.26µs ±10%  -25.73%  (p=0.009 n=10+10)
SpanWithAttributes_4/NeverSample-4               821ns ±66%     608ns ± 5%     ~     (p=0.859 n=10+9)
SpanWithAttributes_8/AlwaysSample-4             4.65µs ±95%    3.39µs ±12%  -27.16%  (p=0.035 n=10+10)
SpanWithAttributes_8/NeverSample-4              1.09µs ±84%    0.81µs ± 7%     ~     (p=0.529 n=10+10)
SpanWithAttributes_all/AlwaysSample-4           3.55µs ±69%    2.37µs ± 6%  -33.25%  (p=0.000 n=10+9)
SpanWithAttributes_all/NeverSample-4             895ns ±74%     661ns ± 8%     ~     (p=0.123 n=10+10)
SpanWithAttributes_all_2x/AlwaysSample-4        5.05µs ±53%    4.28µs ± 6%  -15.37%  (p=0.031 n=9+9)
SpanWithAttributes_all_2x/NeverSample-4         1.17µs ±80%    0.88µs ±10%     ~     (p=0.123 n=10+10)
SpanWithEvents_4/AlwaysSample-4                 3.18µs ±62%    2.44µs ± 7%  -23.32%  (p=0.000 n=10+10)
SpanWithEvents_4/NeverSample-4                   495ns ±34%     429ns ±11%  -13.35%  (p=0.010 n=9+10)
SpanWithEvents_8/AlwaysSample-4                 4.93µs ±64%    3.63µs ± 5%  -26.51%  (p=0.000 n=10+10)
SpanWithEvents_8/NeverSample-4                   554ns ±49%     431ns ± 3%     ~     (p=0.053 n=10+9)
SpanWithEvents_WithStackTrace/AlwaysSample-4    2.27µs ±71%    1.46µs ±11%  -35.33%  (p=0.000 n=10+10)
SpanWithEvents_WithStackTrace/NeverSample-4      640ns ±69%     466ns ± 7%  -27.23%  (p=0.019 n=10+10)
SpanWithEvents_WithTimestamp/AlwaysSample-4     2.11µs ±72%    1.46µs ± 6%  -30.92%  (p=0.000 n=10+9)
SpanWithEvents_WithTimestamp/NeverSample-4       647ns ±57%     535ns ± 4%     ~     (p=0.053 n=10+9)
TraceID_DotString-4                              119ns ±78%     100ns ± 5%     ~     (p=0.781 n=10+10)
SpanID_DotString-4                              73.5ns ± 4%    70.3ns ± 5%   -4.36%  (p=0.012 n=8+10)

name                                          old alloc/op   new alloc/op   delta
SpanProcessor-4                                 12.6kB ± 0%    12.3kB ± 0%   -2.53%  (p=0.000 n=9+10)
SpanProcessorVerboseLogging-4                   13.7kB ± 0%    13.4kB ± 0%   -2.33%  (p=0.000 n=8+9)
StartEndSpan/AlwaysSample-4                       848B ± 0%      816B ± 0%   -3.77%  (p=0.000 n=10+10)
StartEndSpan/NeverSample-4                        224B ± 0%      224B ± 0%     ~     (all equal)
SpanWithAttributes_4/AlwaysSample-4             1.76kB ± 0%    1.73kB ± 0%   -1.82%  (p=0.000 n=10+10)
SpanWithAttributes_4/NeverSample-4                480B ± 0%      480B ± 0%     ~     (all equal)
SpanWithAttributes_8/AlwaysSample-4             2.46kB ± 0%    2.43kB ± 0%   -1.30%  (p=0.000 n=10+10)
SpanWithAttributes_8/NeverSample-4                736B ± 0%      736B ± 0%     ~     (all equal)
SpanWithAttributes_all/AlwaysSample-4           1.94kB ± 0%    1.90kB ± 0%   -1.65%  (p=0.000 n=10+10)
SpanWithAttributes_all/NeverSample-4              544B ± 0%      544B ± 0%     ~     (all equal)
SpanWithAttributes_all_2x/AlwaysSample-4        3.24kB ± 0%    3.20kB ± 0%   -0.99%  (p=0.000 n=10+10)
SpanWithAttributes_all_2x/NeverSample-4           864B ± 0%      864B ± 0%     ~     (all equal)
SpanWithEvents_4/AlwaysSample-4                 1.54kB ± 0%    1.50kB ± 0%   -2.08%  (p=0.000 n=10+10)
SpanWithEvents_4/NeverSample-4                    224B ± 0%      224B ± 0%     ~     (all equal)
SpanWithEvents_8/AlwaysSample-4                 2.24kB ± 0%    2.21kB ± 0%   -1.43%  (p=0.000 n=10+10)
SpanWithEvents_8/NeverSample-4                    224B ± 0%      224B ± 0%     ~     (all equal)
SpanWithEvents_WithStackTrace/AlwaysSample-4    1.02kB ± 0%    0.99kB ± 0%   -3.12%  (p=0.000 n=10+10)
SpanWithEvents_WithStackTrace/NeverSample-4       240B ± 0%      240B ± 0%     ~     (all equal)
SpanWithEvents_WithTimestamp/AlwaysSample-4     1.05kB ± 0%    1.02kB ± 0%   -3.05%  (p=0.000 n=10+10)
SpanWithEvents_WithTimestamp/NeverSample-4        264B ± 0%      264B ± 0%     ~     (all equal)

name                                          old allocs/op  new allocs/op  delta
SpanProcessor-4                                    105 ± 0%        85 ± 0%  -19.05%  (p=0.000 n=10+10)
SpanProcessorVerboseLogging-4                      111 ± 1%        92 ± 1%  -17.86%  (p=0.000 n=10+10)
StartEndSpan/AlwaysSample-4                       9.00 ± 0%      7.00 ± 0%  -22.22%  (p=0.000 n=10+10)
StartEndSpan/NeverSample-4                        3.00 ± 0%      3.00 ± 0%     ~     (all equal)
SpanWithAttributes_4/AlwaysSample-4               19.0 ± 0%      17.0 ± 0%  -10.53%  (p=0.000 n=10+10)
SpanWithAttributes_4/NeverSample-4                4.00 ± 0%      4.00 ± 0%     ~     (all equal)
SpanWithAttributes_8/AlwaysSample-4               27.0 ± 0%      25.0 ± 0%   -7.41%  (p=0.000 n=10+10)
SpanWithAttributes_8/NeverSample-4                4.00 ± 0%      4.00 ± 0%     ~     (all equal)
SpanWithAttributes_all/AlwaysSample-4             21.0 ± 0%      19.0 ± 0%   -9.52%  (p=0.000 n=10+10)
SpanWithAttributes_all/NeverSample-4              4.00 ± 0%      4.00 ± 0%     ~     (all equal)
SpanWithAttributes_all_2x/AlwaysSample-4          32.0 ± 0%      30.0 ± 0%   -6.25%  (p=0.000 n=10+10)
SpanWithAttributes_all_2x/NeverSample-4           4.00 ± 0%      4.00 ± 0%     ~     (all equal)
SpanWithEvents_4/AlwaysSample-4                   20.0 ± 0%      18.0 ± 0%  -10.00%  (p=0.000 n=10+10)
SpanWithEvents_4/NeverSample-4                    3.00 ± 0%      3.00 ± 0%     ~     (all equal)
SpanWithEvents_8/AlwaysSample-4                   29.0 ± 0%      27.0 ± 0%   -6.90%  (p=0.000 n=10+10)
SpanWithEvents_8/NeverSample-4                    3.00 ± 0%      3.00 ± 0%     ~     (all equal)
SpanWithEvents_WithStackTrace/AlwaysSample-4      13.0 ± 0%      11.0 ± 0%  -15.38%  (p=0.000 n=10+10)
SpanWithEvents_WithStackTrace/NeverSample-4       4.00 ± 0%      4.00 ± 0%     ~     (all equal)
SpanWithEvents_WithTimestamp/AlwaysSample-4       14.0 ± 0%      12.0 ± 0%  -14.29%  (p=0.000 n=10+10)
SpanWithEvents_WithTimestamp/NeverSample-4        5.00 ± 0%      5.00 ± 0%     ~     (all equal)
```